Improvements.

+++ data/business/schema.js
@@ -65,7 +65,7 @@ export const schema = [`
     _index: String!
     _type: String!
     _id: String!
-    _score: Int!
+    _score: Float!
     _source: ESUserSource!
   }
